jobBeta3/py/requirements.txt
mysql-connector-python
PyPDF2
google-generativeai